Fix tests of the method `adapter.log`
